Fix punctuation for subwords in ElectraPunctuator
We had the wrong assumption, and didn't have a test, for what the predicted
punctuation looks like when it follows a word that is split into subwords. The
prediction for *each* subword is the predicted punctuation for the whole
word. Using our previous assumption we end up with:

    þú ert mitt sólsk, ##in, mitt eina sólsk. ##in. þú gleð ##ur mig þegar
    heimilin grá ##nar

Which is obviously not correct.