Patch libgomp included with libtorch
We depend on a precompiled and packaged libtorch, which includes the libraries
it depends on. The SONAME of the included libgomp is incorrect, instead of being
libgomp-75eea7e8.so.1 it is libgomp.so.1. Thus, linking to libgomp-75eea7e8.so.1
causes a link to libgomp.so.1. This commit adds a genrule that uses `patchelf`
to fix the soname.